fix(skilleval): polish round 2 — majority vote float + 테스트 2건 추가
- majority vote를 float64 비교로 변경 (tie 보수적 처리 명시)
- TestRunEval_AllTrialsFail: 모든 LLM 호출 실패 시 케이스 집계 계약 검증
- TestRunEval_KClamp: k=200 → 100 클램프 + 100 trials 검증
- mockProvider에 err 필드 추가 (에러 주입 지원)

+++ c4-core/internal/skilleval/runner.go
@@ -182,7 +182,8 @@ func RunEval(ctx context.Context, gateway *llm.Gateway, projectRoot, skillName s
 		cr.AvgConfidence = totalConf / float64(successCount)
 
 		// Majority vote uses only successful calls.
-		majority := trueCount > successCount/2
+		// float comparison avoids integer-division bias; ties (exactly 50%) resolve to false (conservative).
+		majority := float64(trueCount) > float64(successCount)/2.0
 		cr.Correct = majority == test.ShouldTrigger
 
 		// pass@k: at least one trial is correct
